fixes in movement and attack feedback

+++ Assets/Scripts/Data/EnemySettingsSO.cs
@@ -5,11 +5,11 @@ public class EnemySettingsSO : ScriptableObject
 {
     [SerializeField] private float speedMovement;
     [SerializeField] private int damage;
-    [SerializeField] private int timeStun;
+    [SerializeField] private float timeStun;
     [SerializeField] private int timeMoveSound = 0;
 
     public float SpeedMovement { get { return speedMovement; } }
     public int Damage { get { return damage; } }
-    public int TimeStun { get { return timeStun; } }
+    public float  TimeStun { get { return timeStun; } }
     public int TimeMoveSound { get { return timeMoveSound; } }
 }
